Rename models.py to schema.py
